[FEATURE] Merge nginx.locations from data bags
Previously, the `locations` parameter of the `nginx_conf_file`
resource was unused. This limited the flexibility in the data
bag items, as we were not able to specify options for `/`.

Now, the defaults from `node['nginx_conf']['locations']` (defined
in `attributes/nginx.rb` get deep-merged with those specified in
data bag below nginx.locations:

```json
{
  "id": "websockets.example.org",
  "name": "websockets.example.org",
  "nginx": {
    "backend": "http://192.0.2.1:80",
    "locations": {
      "/": {
        "proxy_http_version": "1.1",
        "proxy_set_header": {
          "Upgrade": "$http_upgrade",
          "Connection": "\"upgrade\""
        }
      }
    }
  }
}

```